ui: Move TimelineItem, TimelineItemKind into new sub-module
